requirements.txt
streamlit
pandas
numpy
scikit-learn
xgboost
faiss-cpu
pydeck
plotly